Update schedule.cpp
Made a few changes, first make it so that instead of distributing the processes into different queues, the processes are added to the first queue, and each is executed once, following the algorithm, after that, if a process still has some burst time, then sent that process to the next queue, and continue as mentioned before. If FCFS or Shortest job first or priority, then execute each program a time of burst divided by total number of queues, else if round robin, then just execute once if burst time greater than quantum time, otherwise send to next queue. For the last queue, just follow the normal rules, and normal algorithm.

The rest remained similar.